# src/main/java/org/openapitools/api/AlbumsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Album;
import org.openapitools.model.UploadTrackRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-03T18:02:44.513930+01:00[Europe/Madrid]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "contents", description = "Operaciones de artistas, discográficas, álbumes, canciones y archivos de audio.")
public interface AlbumsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_ALBUM_BY_ID = "/albums/{idAlbum}";
    /**
     * GET /albums/{idAlbum} : Información de álbum por ID
     * Obtiene los datos completos de un álbum específico.
     *
     * @param idAlbum  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "getAlbumById",
        summary = "Información de álbum por ID",
        description = "Obtiene los datos completos de un álbum específico.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Album.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = AlbumsApi.PATH_GET_ALBUM_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<Album> getAlbumById(
        @NotNull @Parameter(name = "idAlbum", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idAlbum") String idAlbum
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cover_url\" : \"https://avatar.iran.liara.run/public/40\", \"release_date\" : \"2008-09-09\", \"id\" : \"alb_101\", \"title\" : \"La ley innata\", \"artist_id\" : \"art_001\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_ALBUMS = "/albums";
    /**
     * GET /albums : Listado de álbumes
     * Devuelve el listado completo de álbumes publicados.
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "listAlbums",
        summary = "Listado de álbumes",
        description = "Devuelve el listado completo de álbumes publicados.",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Album.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = AlbumsApi.PATH_LIST_ALBUMS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<Album>> listAlbums(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"cover_url\" : \"https://avatar.iran.liara.run/public/40\", \"release_date\" : \"2008-09-09\", \"id\" : \"alb_101\", \"title\" : \"La ley innata\", \"artist_id\" : \"art_001\" }, { \"cover_url\" : \"https://avatar.iran.liara.run/public/40\", \"release_date\" : \"2008-09-09\", \"id\" : \"alb_101\", \"title\" : \"La ley innata\", \"artist_id\" : \"art_001\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPLOAD_TRACK = "/albums/{idAlbum}/tracks";
    /**
     * POST /albums/{idAlbum}/tracks : Subir canción
     * Añade una nueva canción a un álbum (requiere rol artista).
     *
     * @param idAlbum  (required)
     * @param uploadTrackRequest Datos de la canción a subir (required)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     *         or Not Found (status code 404)
     */
    @Operation(
        operationId = "uploadTrack",
        summary = "Subir canción",
        description = "Añade una nueva canción a un álbum (requiere rol artista).",
        tags = { "contents" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created"),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "404", description = "Not Found")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AlbumsApi.PATH_UPLOAD_TRACK,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> uploadTrack(
        @NotNull @Parameter(name = "idAlbum", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idAlbum") String idAlbum,
        @Parameter(name = "UploadTrackRequest", description = "Datos de la canción a subir", required = true) @Valid @RequestBody UploadTrackRequest uploadTrackRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}