# src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  description: "Administración de artistas, discográficas, álbumes, canciones y archivos\
    \ de audio."
  title: Microservicio de Contenidos
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /albums/{idAlbum}:
    get:
      description: Obtiene los datos completos de un álbum específico.
      operationId: getAlbumById
      parameters:
      - explode: false
        in: path
        name: idAlbum
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/album"
          description: OK
        "404":
          description: Not Found
      summary: Información de álbum por ID
      tags:
      - contents
      x-accepts:
      - application/json
      x-tags:
      - tag: contents
...
  /albums/{idAlbum}/tracks:
    post:
      description: Añade una nueva canción a un álbum (requiere rol artista).
      operationId: uploadTrack
      parameters:
      - explode: false
        in: path
        name: idAlbum
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/uploadTrack_request"
        description: Datos de la canción a subir
        required: true
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Subir canción
      tags:
      - contents
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: contents
...
components:
  schemas:
...
    album:
      example:
        cover_url: https://avatar.iran.liara.run/public/40
        release_date: 2008-09-09
        id: alb_101
        title: La ley innata
        artist_id: art_001
      properties:
        id:
          example: alb_101
          type: string
        artist_id:
          example: art_001
          type: string
        title:
          example: La ley innata
          type: string
        release_date:
          example: 2008-09-09
          type: string
        cover_url:
          example: https://avatar.iran.liara.run/public/40
          type: string
      type: object
...
    uploadTrack_request:
      properties:
        title:
          example: De acero
          type: string
        duration:
          example: 03:45
          type: string
        file_url:
          example: https://cdn.musica.com/tracks/de-acero.mp3
          type: string
      required:
      - duration
      - file_url
      - title
      type: object